requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
pillow
pymongo
python-dotenv
insightface
onnxruntime
boto3
